+++ CHANGES.rst
@@ -1,7 +1,14 @@
 Changelog
 =========
 
+1.0.1
+-----
+
+* Added arbitrary script operation (bubenkoff)
+* Strict phases configuration assertions for phased migration decorator (bubenkoff)
+* `get_migration_data` API (bubenkoff)
+
 1.0.0
 -----
 
-* Initial public release
+* Initial public release (bubenkoff)